Fix all Phase 0011 code review findings
Bug 1 (BS001 board scan overcount): Create zero_i32 constant outside
carried_inputs; use as false_val in select to prevent overcounting.

Bug 2 (context_id always ContextId(0)): Add context_id field to
TransformationContext, set during database insert, use in generators.

Bug 3 (add_description overwrite): Add debug_assert for duplicate regions.

Fix 4 (u32 truncation): Change Support fields from u16 to u32; return
i64 from score(). Update all consumers.

Fix 5 (test name): Rename non_bitset_context_produces_no_candidates
to bitmask_context_produces_four_candidates with corrected comments.

Fix 6 (generator dedup): Replace 4 duplicate generator files with a
single data-driven bitset.rs. Delete bit_iteration.rs, popcount.rs,
packed_bitfield.rs, mask_construction.rs.

Fix 7 (CandidateExplanation): Remove redundant strategy, representation,
and prerequisites fields. Keep only source_concepts and rationale.

Fix 8 (iterator return): all_plans() returns impl Iterator<Item=Candidate>
instead of Vec<Candidate>. Update consumers.

Fix 9 (serde): Add Serialize/Deserialize to Representation, SourceStructure,
Constraint, Assumption, ContextId, TransformationContext, CandidateId,
ImplementationStrategy, CandidateEffects, RegionId.

Fix 10 (static assumptions): Extract DEFAULT_ASSUMPTIONS constant in
engine.rs.

+++ sir/crates/sir_inference/src/hypothesis.rs
@@ -3,14 +3,14 @@ pub use sir_transform::representation::Representation;
 /// Integer support score — no floating point in engine logic.
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
 pub struct Support {
-    pub positive: u16,
-    pub negative: u16,
+    pub positive: u32,
+    pub negative: u32,
 }
 
 impl Support {
     /// Net score: positive minus negative.
-    pub fn score(&self) -> i32 {
-        self.positive as i32 - self.negative as i32
+    pub fn score(&self) -> i64 {
+        self.positive as i64 - self.negative as i64
     }
 
     /// Ratio of positive support to total (for display only).
